ConstitutiveModels

Documentation for ConstitutiveModels.

ConstitutiveModels.AbstractMotion

ConstitutiveModels.HardeningModel

ConstitutiveModels.IsochoricUniaxialStress

ConstitutiveModels.IsotropicHardeningModel

ConstitutiveModels.LinearIsotropicHardening

ConstitutiveModels.LinearIsotropicHardening

ConstitutiveModels.NoIsotropicHardening

ConstitutiveModels.NoIsotropicHardening

ConstitutiveModels.PlasticityModel

ConstitutiveModels.SimpleMotion

ConstitutiveModels.SimpleShear

ConstitutiveModels.UniaxialStrain

ConstitutiveModels.deformation_gradient

ConstitutiveModels.deformation_gradient

ConstitutiveModels.deformation_gradient

ConstitutiveModels.deformation_gradient

ConstitutiveModels.deformation_gradient

ConstitutiveModels.elastic_properties

ConstitutiveModels.radius

ConstitutiveModels.radius

ConstitutiveModels.slope

ConstitutiveModels.slope

ConstitutiveModels.update

ConstitutiveModels.update

ConstitutiveModels.yield_stress

ConstitutiveModels.AbstractMotion — Type

ConstitutiveModels.HardeningModel — Type

ConstitutiveModels.IsochoricUniaxialStress — Type

Provides an analytic motion for uniaxial stress assuming perfect incompressibility.

This is

\[\mathbf{F} = \begin{bmatrix} \lambda & 0 & 0 \\ 0 & \frac{1}{\sqrt{\lambda}} & 0 \\ 0 & 0 & \frac{1}{\sqrt{\lambda}} \end{bmatrix}\]

ConstitutiveModels.IsotropicHardeningModel — Type

ConstitutiveModels.LinearIsotropicHardening — Type

ConstitutiveModels.LinearIsotropicHardening — Method

LinearIsotropicHardening(
    inputs::Dict
) -> Tuple{ConstitutiveModels.LinearIsotropicHardening, Any, StaticArraysCore.SVector{0, Float64}}

LinearIsotropicHardening(inputs)

defined at /home/runner/work/ConstitutiveModels.jl/ConstitutiveModels.jl/src/plasticity_models/HardeningModels.jl:80.

ConstitutiveModels.NoIsotropicHardening — Type

ConstitutiveModels.NoIsotropicHardening — Method

NoIsotropicHardening(
    _::Dict
) -> Tuple{ConstitutiveModels.NoIsotropicHardening, StaticArraysCore.SVector{0, Float64}, StaticArraysCore.SVector{0, Float64}}

NoIsotropicHardening(_)

defined at /home/runner/work/ConstitutiveModels.jl/ConstitutiveModels.jl/src/plasticity_models/HardeningModels.jl:44.

ConstitutiveModels.PlasticityModel — Type

ConstitutiveModels.SimpleMotion — Type

ConstitutiveModels.SimpleShear — Type

Provides an analytic motion for simple shear.

This is

\[\mathbf{F} = \begin{bmatrix} 1 & \gamma & 0 \\ 0 & 1 & 0 \\ 0 & 0 & 1 \end{bmatrix}\]

ConstitutiveModels.UniaxialStrain — Type

Provides an analytic motion for uniaxial strain

This is

\[\mathbf{F} = \begin{bmatrix} \lambda & 0 & 0 \\ 0 & 1 & 0 \\ 0 & 0 & 1 \end{bmatrix}\]

ConstitutiveModels.deformation_gradient — Function

Returns the deformation gradient for a given motion

deformation_gradient(_, λ, type)

defined at /home/runner/work/ConstitutiveModels.jl/ConstitutiveModels.jl/src/Motions.jl:31.

deformation_gradient(_, λ, type)

defined at /home/runner/work/ConstitutiveModels.jl/ConstitutiveModels.jl/src/Motions.jl:33.

deformation_gradient(_, λ, type)

defined at /home/runner/work/ConstitutiveModels.jl/ConstitutiveModels.jl/src/Motions.jl:35.

deformation_gradient(_, λ, type)

defined at /home/runner/work/ConstitutiveModels.jl/ConstitutiveModels.jl/src/Motions.jl:37.

deformation_gradient(_, γ, _)

defined at /home/runner/work/ConstitutiveModels.jl/ConstitutiveModels.jl/src/Motions.jl:57.

deformation_gradient(_, γ, _)

defined at /home/runner/work/ConstitutiveModels.jl/ConstitutiveModels.jl/src/Motions.jl:59.

deformation_gradient(_, γ, _)

defined at /home/runner/work/ConstitutiveModels.jl/ConstitutiveModels.jl/src/Motions.jl:61.

deformation_gradient(_, γ, type)

defined at /home/runner/work/ConstitutiveModels.jl/ConstitutiveModels.jl/src/Motions.jl:63.

deformation_gradient(_, λ, _)

defined at /home/runner/work/ConstitutiveModels.jl/ConstitutiveModels.jl/src/Motions.jl:83.

deformation_gradient(_, λ, _)

defined at /home/runner/work/ConstitutiveModels.jl/ConstitutiveModels.jl/src/Motions.jl:85.

deformation_gradient(_, λ, _)

defined at /home/runner/work/ConstitutiveModels.jl/ConstitutiveModels.jl/src/Motions.jl:87.

deformation_gradient(_, λ, type)

defined at /home/runner/work/ConstitutiveModels.jl/ConstitutiveModels.jl/src/Motions.jl:89.

deformation_gradient(motion, model, props, state, λ, type)

defined at /home/runner/work/ConstitutiveModels.jl/ConstitutiveModels.jl/src/Motions.jl:160.

deformation_gradient(motion, λ; type)

defined at /home/runner/work/ConstitutiveModels.jl/ConstitutiveModels.jl/src/Motions.jl:189.

deformation_gradient(motion, model, props, state, λ; type)

defined at /home/runner/work/ConstitutiveModels.jl/ConstitutiveModels.jl/src/Motions.jl:191.

ConstitutiveModels.deformation_gradient — Method

deformation_gradient(
    motion::Type{UniaxialStressDisplacementControl},
    model::ConstitutiveModel,
    props::AbstractArray{T<:Number, 1},
    state,
    λ::Number,
    type::Type
) -> Any

deformation_gradient(motion, model, props, state, λ, type)

defined at /home/runner/work/ConstitutiveModels.jl/ConstitutiveModels.jl/src/Motions.jl:160.

ConstitutiveModels.deformation_gradient — Method

deformation_gradient(
    _::Type{IsochoricUniaxialStress},
    λ::Number,
    type::Type{<:Tensors.Tensor}
) -> Tensors.Tensor{2, 3, _A, 9} where _A

deformation_gradient(_, λ, type)

defined at /home/runner/work/ConstitutiveModels.jl/ConstitutiveModels.jl/src/Motions.jl:31.

ConstitutiveModels.deformation_gradient — Method

deformation_gradient(
    _::Type{SimpleShear},
    γ::Number,
    _::Type{<:Tensors.Tensor}
) -> Tensors.Tensor{2, 3, _A, 9} where _A

deformation_gradient(_, γ, _)

defined at /home/runner/work/ConstitutiveModels.jl/ConstitutiveModels.jl/src/Motions.jl:57.

ConstitutiveModels.deformation_gradient — Method

deformation_gradient(
    _::Type{UniaxialStrain},
    λ::Number,
    _::Type{<:Tensors.Tensor}
) -> Tensors.Tensor{2, 3, _A, 9} where _A

deformation_gradient(_, λ, _)

defined at /home/runner/work/ConstitutiveModels.jl/ConstitutiveModels.jl/src/Motions.jl:83.

ConstitutiveModels.elastic_properties — Method

elastic_properties(props::AbstractVector{<:Number}) -> Any

elastic_properties(props)

defined at /home/runner/work/ConstitutiveModels.jl/ConstitutiveModels.jl/src/MechanicalModels.jl:47.

ConstitutiveModels.radius — Method

radius(
    _::ConstitutiveModels.LinearIsotropicHardening,
    props,
    eqps
) -> Any

radius(_, props, eqps)

defined at /home/runner/work/ConstitutiveModels.jl/ConstitutiveModels.jl/src/plasticity_models/HardeningModels.jl:88.

ConstitutiveModels.radius — Method

radius(
    _::ConstitutiveModels.NoIsotropicHardening,
    props::AbstractArray,
    eqps
) -> Any

radius(_, props, eqps)

defined at /home/runner/work/ConstitutiveModels.jl/ConstitutiveModels.jl/src/plasticity_models/HardeningModels.jl:50.

ConstitutiveModels.slope — Method

slope(
    _::ConstitutiveModels.LinearIsotropicHardening,
    props,
    eqps
) -> Any

slope(_, props, eqps)

defined at /home/runner/work/ConstitutiveModels.jl/ConstitutiveModels.jl/src/plasticity_models/HardeningModels.jl:91.

ConstitutiveModels.slope — Method

slope(
    _::ConstitutiveModels.NoIsotropicHardening,
    props::AbstractArray,
    eqps
) -> Float64

slope(_, props, eqps)

defined at /home/runner/work/ConstitutiveModels.jl/ConstitutiveModels.jl/src/plasticity_models/HardeningModels.jl:53.

ConstitutiveModels.update — Method

update(
    _::ConstitutiveModels.LinearIsotropicHardening,
    yield_surf::ConstitutiveModels.J2YieldSurface,
    props,
    μ,
    σ,
    α_old
) -> Any

update(_, yield_surf, props, μ, σ, α_old)

defined at /home/runner/work/ConstitutiveModels.jl/ConstitutiveModels.jl/src/plasticity_models/HardeningModels.jl:95.

ConstitutiveModels.update — Method

update(
    _::ConstitutiveModels.NoIsotropicHardening,
    yield_surf::ConstitutiveModels.J2YieldSurface,
    props,
    μ,
    σ,
    α_old
) -> Any

Specialized method for update so no non-linear solve takes place

update(_, yield_surf, props, μ, σ, α_old)

defined at /home/runner/work/ConstitutiveModels.jl/ConstitutiveModels.jl/src/plasticity_models/HardeningModels.jl:57.

ConstitutiveModels.yield_stress — Method

yield_stress(
    _::ConstitutiveModels.YieldSurface,
    props::AbstractArray
) -> Any

Always assumes the yield stress is the first prop

yield_stress(_, props)

defined at /home/runner/work/ConstitutiveModels.jl/ConstitutiveModels.jl/src/plasticity_models/YieldSurfaces.jl:6.