ConstitutiveModels.initialize_props — Method

All models are assumed to take in a type tag for the model and a Dict of props

This method we likely can't statically compile due to the Any

ConstitutiveModels.initialize_state — Method

Default state constructor to just return zeros